+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 ## Unreleased
 
+FEATURES:
+* Now when calling RemoveInstance, if an empty replicaset name is passed, the replicaset will be calculated automatically.
+
 CHANGES:
 * Bump go-tarantool from v2.2.1 to v2.3.0.
 
 TESTS:
 * Fixed etcd overlapping ports.
+* Fixed Pooler mocks generation.
 
 ## v2.0.4